requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Werkzeug==3.0.3
SQLAlchemy==2.0.36
reportlab==4.2.2
python-socketio==5.11.4
Pillow==10.4.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
eventlet==0.35.2
python-dotenv==1.0.0
